STAC-8617: Work out molecule tests
