+++ deployment/kubernetes/test_connections/pod-localhost.yaml
@@ -6,7 +6,7 @@ spec:
   containers:
     - image: jgkamat/netcat
       name: server
-      command: ["/bin/bash", "-c", "nc -l -p 123 > /dev/null"]
+      command: ["/bin/bash", "-c", "nc -l -p 234 > /dev/null"]
     - image: jgkamat/netcat
       name: client
-      command: ["/bin/bash", "-c", "yes | nc localhost 123"]
+      command: ["/bin/bash", "-c", "yes | nc localhost 234"]
