+++ stackstate-deps.json
@@ -4,7 +4,7 @@
   "TRACE_AGENT_BRANCH": "6.7.11",
   "TRACE_AGENT_WINDOWS_BINARY": "yes",
   "PROCESS_AGENT_BRANCH": "stac-8617",
-  "PROCESS_AGENT_VERSION": "2.6.0-480d3e6",
+  "PROCESS_AGENT_VERSION": "2.6.0-bb19556",
   "OMNIBUS_SOFTWARE_VERSION": "6.11.0",
   "OMNIBUS_RUBY_VERSION": "05fb405a100b8157bc5b32e48a97e674857e97f6",
   "JMXFETCH_VERSION": "0.26.1",
