+++ test/molecule-role/molecule/kubernetes/group_vars/all.yml
@@ -11,4 +11,6 @@ stackstate_branch: "{{ lookup('env', 'STACKSTATE_BRANCH') }}"
 cluster_name: "molecule_k8s_test"
 dnat_service_port: 321
 dnat_server_port: 123
+container_to_container_server_port: 234
+headless_service_port: 543
 namespace: "{{ agent_current_branch.replace('.','-') | lower }}"
